Changing the docker compose runner

+++ run_safe.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 git checkout feature/ts_support
-cp .env.example .env
-cp api/.env.example api/.env
-cp my-app/.env.example my-app/.env
+
+if [ ! -f .env ]; then
+    echo "WARNING: using .env.example to populate .env: better if you create your own .env file"
+    cp .env.example .env
+fi
+
+echo "NOTICE: Changing the permissions of the database files to make sure the docker container can copy them"
 chmod 777 db/*
 docker compose -f build.docker-compose.yml up
